docs(readme):Fatjar Link Outdated
I was trying to run jikkou using a Fatjar and ran into this link. After browsing for a bit I noticed the link was wrong, as well as other things described in the issue. For this MR I am only making a small change to the README file not fixing the upload to Maven.

Issue: https://github.com/streamthoughts/jikkou/issues/302